% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\docType{data}
\name{crosswalk}
\alias{crosswalk}
\title{Crosswalk between abbreviations and species names for North
     American epiphytic macrolichens}
\format{An object of class \code{data.frame} with 470 rows and 2 columns.}
\usage{
data(crosswalk)
}
\description{
This data set gives information about FIA abbreviations and the
     corresponding species names for North American macrolichens.
}
\details{
Species data were sourced from the US Forest Service's
     Forest Inventory and Analysis program (FIA 2011).
}
\references{
US Forest Service, Forest Inventory and Analysis [FIA]. 2011. Phase
     3 Field Guide – Lichen Communities, Version 5.1.
}
\keyword{datasets}
