% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_q95.R
\name{fit_q95}
\alias{fit_q95}
\title{Fit 95th percentile as nonlinear 2D surface}
\usage{
fit_q95(data, ...)
}
\arguments{
\item{data}{data.frame of observations, with columns named
\code{x1}, \code{x2}, and \code{y}.}

\item{...}{further arguments to \code{persp}.}

\item{m}{fitted model from \code{crs}.}

\item{zlo, zhi}{upper and lower z-limits for plotting.}

\item{zlab}{z-axis label.}

\item{txt}{character, text to add as plot label.}
}
\value{
\code{fit_q95} returns a fitted \code{crs} object, \code{persp_q95}
    plots its perspective plot, and \code{qlab} is a convenience
    labeling function.
}
\description{
Hard-coded to fit nonlinear regression surface.
}
\details{
Hard-coded, so only for one custom plotting purpose.
}
