% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatindex.R
\name{heatindex}
\alias{heatindex}
\title{Heat index}
\usage{
heatindex(TC, RH, round = 3)
}
\arguments{
\item{TC}{vector of temperature values (in Celsius)}

\item{RH}{vector of relative humidity values (as a percentage
[0,100])}

\item{...}{further arguments passed to other functions.}
}
\value{
vector of heat index values, in Celsius.
}
\description{
Heat index based on temperature and relative humidity. Useful for
    describing simultaneously hot and moist conditions, such as
    those that may limit positive carbon balance for lichens.
}
\details{
The heat index approximates \sQuote{experienced} heat.
}
\examples{
set.seed(99)
TC <- sort(runif(99, 5, 35)+1)
RH <- sort(runif(99, 50, 90))
par(mar=c(4,4,0,4), pty='s')
plot(RH, TC, ylim=c(0,80))
points(RH, heatindex(TC,RH), col=2)
axis(4, at = seq(0,80,len=5), labels = seq(0,80,len=5), col.axis=2)
mtext('Heat index', 4, line=2, col=2)

}
\seealso{
https://www.wpc.ncep.noaa.gov/html/heatindex_equation.shtml for the
    algorithm, and \link[weathermetrics]{heat.index} for a more
    complicated adaptation.
}
