% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmap.R
\name{mmap}
\alias{mmap}
\title{Regional mapping}
\usage{
mmap(x, field, name = field, title = "", alpha = NA, sizept = 1,
  sizetxt = 1, xlim = c(-122, -75), ylim = c(25, 49.5), bg = NA,
  colorscale, midval = 0.6, ...)
}
\arguments{
\item{x}{data.frame of observations.}

\item{field}{character, name of column in \code{x} by which to
color points.}

\item{name}{character, name for color scale.}

\item{title}{character, main title.}

\item{alpha}{numeric vector, values used for relative alpha
transparency.}

\item{sizept, sizetxt}{numeric, relative size for points and text.}

\item{xlim}{numeric vector, x limits (longitude in decimal
degrees).}

\item{ylim}{numeric vector, y limits (latitude in decimal degrees).}

\item{bg}{character, background fill for land area (defaults to
transparent).}

\item{colorscale}{character, one of \code{c('bw','bw2','inferno',
'discretebw')} defining the color scale.}

\item{midval}{numeric, value at color scale midpoint when
colorscale = 'bw2'.}

\item{...}{further arguments passed to ggplot2 color scales (for
example, transformations).}
}
\value{
Returns or plots an object inheriting from class \code{'ggplot'}.
}
\description{
Map geographic locations as points colored by values.
}
\details{
Plot continuous or discrete values mapped across regions of North
    America.
}
\examples{
# iris data
set.seed(1959)
x <- iris[,1:3]
n <- NROW(x)
# add geographic structure
x <- data.frame(
     x,
     lon = seq(-80,-120,len=n) + rnorm(n,0,2),
     lat = seq(30,45,len=n) + rnorm(n,0,2),
     Sepal.Big = as.factor(ifelse(x$Sepal.Length > 6,1,0)))
(mmap(x, 'Sepal.Length', alpha=1, name='Sepal length'))
(mmap(x, 'Sepal.Length', alpha=1, name='Sepal length',
      colorscale='inferno'))
(mmap(x, 'Sepal.Big',  name='Sepal > 6 mm'))

}
