% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{noise}
\alias{noise}
\title{Noise}
\usage{
noise(z, perc = 1e-04, unif = TRUE, ...)
}
\arguments{
\item{z}{vector, matrix, or data.frame of observations.}

\item{perc}{numeric, percent of data range.}

\item{unif}{logical, use default uniform distribution? Alternative
is to use normal distribution.}

\item{...}{further arguments passed to other functions.}
}
\value{
Matrix of noisified data of same dimensions as \code{'z'}.
}
\description{
Add random uncertainty for sensitivity analysis.
}
\details{
Adds noise, ignoring NA values.  If \code{unif=TRUE}, then
     uniformly-distributed noise is bounded within \code{perc}
     percent of the original data range.  If \code{unif=FALSE},
     then normally-distributed noise has standard deviation equal
     to \code{perc} percent of the original data range.  Useful for
     perturbing values for sensitivity analysis.
}
\examples{
# iris data
x  <- iris[,1:3]
xn <- noise(x, perc=1)
pairs(x, lower.panel=NULL)
pairs(xn, lower.panel=NULL)

}
