% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens}
\alias{sens}
\title{Sensitivity}
\usage{
sens(xvec, ybin, perc, nrep, ...)
}
\arguments{
\item{xvec}{vector of climate values.}

\item{ybin}{data.frame, of class \code{'bingrid'}..}

\item{perc}{numeric, amount of uniformly distributed noise to add,
expressed as a percentage of the range of \code{xvec}}

\item{nrep}{integer, number of replicates.}

\item{...}{further arguments passed to other functions.}
}
\value{
data.frame of class \code{'sens'}.
}
\description{
Sensitivity of vulnerability outputs to uncertainty in climate
   variable inputs.
}
\details{
Evaluates effects of uncertainty in climate inputs on the
    vulnerability values.
}
\seealso{
\link[vuln]{utils} for internal functions.
}
