% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{make_d}
\alias{make_d}
\alias{plot_d}
\alias{txt_tau}
\alias{txt_val}
\alias{set_par}
\alias{add_text}
\alias{colvec}
\alias{surfcol}
\alias{plot_loess}
\title{Utility plotting functions}
\usage{
make_d(x, ...)

plot_d(d, plotd = FALSE, ...)

txt_tau(e, tau = 0.95, ht, i, ...)

txt_val(e1, e2, e3, xval = 18.5, ht, ...)

set_par(panels = 1, ...)

add_text(x, y, labels, bold = FALSE, ...)

colvec(x, n = 99, alpha = 0.6, begin = 0.2, end = 0.9, dir = 1,
  pal, ...)

surfcol(x, ngrid, alpha = 0.6, begin = 0.2, end = 0.9, dir = 1,
  pal, ...)

plot_loess(x, y, col = "#00000040", lcol = "#FF0000BF", cex = 0.8,
  pch = 16, las = 1, bty = "L", ...)
}
\arguments{
\item{x}{vector of climate values at which a species occurs.}

\item{...}{further arguments passed to other functions.}

\item{d}{density object.}

\item{plotd}{logical, plot the density as polygon?}

\item{e}{ECDF object.}

\item{tau}{specified quantile level.}

\item{ht}{height of text label in relative coordinates.}

\item{i}{character, text in subscript.}

\item{e1, e2, e3}{three ECDF objects.}

\item{xval}{specified x value to return quantile values.}

\item{panels}{number of figure panels in `set_par`.}

\item{y}{numeric, relative vertical position of text in `add_text`.}

\item{labels}{character vector of text to plot in `add_text`.}

\item{bold}{logical, should panel label be bold in `add_text`?}

\item{n}{number of unique colors to be used in the palette.}

\item{alpha}{alpha transparency in [0,1].}

\item{begin}{hue in [0,1] at which the viridis colormap begins.}

\item{end}{hue in [0,1] at which the viridis colormap ends.}

\item{dir}{order of colors in the viridis colormap scale; if 1,
the default, colors are ordered from darkest to lightest; if
-1, then reversed.}

\item{pal}{optional specification for a color palette; defaults to
\code{viridis::inferno} palette if missing.}

\item{ngrid}{number of grid cells in the matrix of values plotted
in `persp`.}

\item{col, lcol, cex, pch, las, bty}{graphical parameters following
\code{\link[graphics]{par}} and \code{\link[graphics]{plot}}.}
}
\value{
See below.
}
\description{
Utility functions for plotting conceptual figures.
}
\details{
Lightweight functions to plot conceptual figures.
}
