% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuln.R
\name{vuln}
\alias{vuln}
\alias{gv}
\title{Multidimensional vulnerability index}
\usage{
vuln(spe, y, ybin, yinv, na.rm = TRUE, fullinfo = TRUE, ...)

gv(obj, pick = 1, ...)
}
\arguments{
\item{spe}{species abundance or occurrence matrix (response
variables).}

\item{y}{vector, matrix, or list of explanatory variables (climate,
fire, etc), where rows correspond to rows in \code{spe}.}

\item{ybin}{optional, data.frame or list of data.frames from
\code{bingrid}, with binned explanatory variables from
rasterizing.  Function gives extended niches if supplied,
possibly truncated niches if not supplied.}

\item{yinv}{vector of logicals, in order of \code{y}, defining
which explanatory variables y should be `inverted` (e.g., low
moisture = high vulnerability). Defaults to
\code{rep(FALSE, NCOL(y))}.}

\item{na.rm}{logical, indicating whether NA values should be
stripped before computation.}

\item{fullinfo}{logical, default \code{fullinfo=TRUE} propagates
any \code{NA} cases that occur in any \code{y} variable;
when \code{fullinfo=TRUE}, percentiles will be based only on
non-NA \code{y} cases.}

\item{...}{further arguments passed to other functions.}

\item{obj}{object of class \code{'vuln'}.}

\item{pick}{numeric, indicating which vulnerability metric (1,2,3).}
}
\value{
Object of class \code{'vuln'}.
}
\description{
Multidimensional vulnerability index, using truncated or extended
    niches.
}
\examples{
\dontrun{
# VI from 'truncated' niches based on point FIA data
v1 <- vuln(spe, y=id$mwmt, na.rm=T)
v2 <- vuln(spe, y=id$cmd,  na.rm=T)
v3 <- vuln(spe, y=id$fire, na.rm=T)
v4 <- vuln(spe, y = cbind(mwmt=id$mwmt,cmd=id$cmd,fire=id$fire),
           na.rm = T)
# VI from 'extended' niches based on rasterized CNALH herb. data
v1 <- vuln(spe, y=id$mwmt,  ybin=m_mwmt, na.rm=T)
v2 <- vuln(spe, y=id$cmd,   ybin=m_cmd,  na.rm=T)
v3 <- vuln(spe, y=id$fire,  ybin=m_fire, na.rm=T)
v4 <- vuln(spe,
           y = cbind(mwmt=id$mwmt,cmd=id$cmd,fire=id$fire),
           ybin = list(mwmt=m_mwmt, cmd=m_cmd, fire=m_fire),
           na.rm = T)
plot(gv(v1,1), gv(v2,1))
}
}
