% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmap.R
\name{mmap}
\alias{mmap}
\alias{dmap}
\title{Custom mapping functions}
\usage{
mmap(x, ptcol, name = ptcol, title = "", alpha = NULL, size = 1,
  xlim = c(-122, -75), ylim = c(25, 49.5), tt = F, bg = NA,
  low = "white", high = "black", ...)

dmap(x, ptcol, name = ptcol, title = "", alpha = NULL, size = 1,
  xlim = c(-122, -75), ylim = c(25, 49.5), bg = NA, ...)
}
\arguments{
\item{x}{data.frame of observations.}

\item{ptcol}{character, name of column in \code{x} by which to
color points.}

\item{name}{character, name for color scale.}

\item{title}{character, main title.}

\item{alpha}{numeric vector, values used for relative alpha
transparency.}

\item{size}{numeric, point size.}

\item{xlim}{numeric vector, x limits (longitude).}

\item{ylim}{numeric vector, x limits (latitude).}

\item{tt}{logical, two-tone color scale?}

\item{bg}{character, background fill for land area (defaults to
transparent).}

\item{low}{character, lower end of color scale.}

\item{high}{character, upper end of color scale.}

\item{...}{further arguments passed to other functions.}
}
\value{
Plots to device, or else object of class \code{'zzz'}.
}
\description{
Map geographic locations as points colored by values.
}
\details{
Plot continuous (\code{'mmap'})or discrete values (\code{'dmap'}).
}
\examples{
# iris data
set.seed(1959)
x <- iris[,1:3]
n <- NROW(x)
# add geographic structure
x <- data.frame(
     x,
     lon = seq(-80,-120,len=n) + rnorm(n,0,2),
     lat = seq(30,45,len=n) + rnorm(n,0,2),
     Sepal.Big = as.factor(ifelse(x$Sepal.Length > 6,1,0)))
(mmap(x, 'Sepal.Length', alpha=1, name='Sepal length'))
(dmap(x, 'Sepal.Big',  name='Sepal > 6 mm'))

}
