% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{make_d}
\alias{make_d}
\alias{plot_d}
\alias{ftxt_tau}
\alias{ftxt_val}
\title{Utility functions for plotting}
\usage{
make_d(x, ...)

plot_d(d, plotd = FALSE, ...)

ftxt_tau(e, tau = 0.95, ht, ...)

ftxt_val(e1, e2, e3, xval = 18.5, ht, ...)
}
\arguments{
\item{x}{vector of climate values at which a species occurs.}

\item{...}{further arguments passed to other functions.}

\item{d}{density object.}

\item{e}{ECDF object.}

\item{tau}{specified quantile level.}

\item{ht}{height of text label in relative coordinates.}

\item{e1, e2, e3}{three ECDF objects.}

\item{xval}{specified x value to return quantile values.}
}
\value{
See below.
}
\description{
Various utility functions.
}
\details{
Lightweight functions to plot.
}
