% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{litetvi}
\alias{litetvi}
\alias{quantileonly}
\alias{percentileonly}
\alias{describe}
\alias{mae}
\title{Utility functions for vulnerability analysis}
\usage{
litetvi(spe, y, ybin, ...)

quantileonly(spe, ybin, na.rm = TRUE, ...)

percentileonly(spe, y, ybin, ...)

describe(x, ...)

mae(x, y, stdz = FALSE, na.rm = TRUE, ...)
}
\arguments{
\item{spe}{matrix or data.frame, species occurrences or abundances.}

\item{y}{vector, environmental values matching \code{ybin}.}

\item{ybin}{data.frame, of class \code{'bingrid'}.}

\item{...}{further arguments passed to other functions.}

\item{na.rm}{logical, should NAs be removed?}
}
\value{
See below.
}
\description{
Various utility functions.
}
\details{
Lightweight functions to calculate vulnerability indices
     (\code{litetvi}), specific quantiles (\code{quantileonly}), or
     all percentiles (\code{percentileonly}).
}
\examples{
### TODO

}
