% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuln.R
\name{vuln}
\alias{vuln}
\alias{gv}
\title{Multidimensional vulnerability index}
\usage{
vuln(spe, y, ybin, yinv, na.rm = TRUE, fullinfo = TRUE, ...)

gv(obj, pick = 1, ...)
}
\arguments{
\item{spe}{species abundance or occurrence matrix (response
variables).}

\item{y}{vector, matrix, or list of explanatory variables (climate,
fire, etc), where rows correspond to rows in \code{spe}.}

\item{ybin}{optional, data.frame or list of data.frames from
\code{bingrid}, with binned explanatory variables from
rasterizing.  Function gives extended niches if supplied,
possibly truncated niches if not supplied.}

\item{yinv}{vector of logicals, in order of \code{y}, defining
which explanatory variables y should be `inverted` (e.g., low
moisture = high vulnerability). Defaults to
\code{rep(FALSE, NCOL(y))}.}

\item{na.rm}{logical, indicating whether to strip NA values before
computation. Default is \code{na.rm = TRUE}}

\item{fullinfo}{logical, default \code{fullinfo=TRUE} propagates
any \code{NA} cases that occur in any \code{y} variable;
when \code{fullinfo=TRUE}, percentiles will be based only on
non-NA \code{y} cases.}

\item{...}{further arguments passed to other functions.}

\item{obj}{object of class \code{'vuln'}.}

\item{pick}{numeric, indicating which vulnerability metric (1,2,3).}
}
\value{
Object of class \code{'vuln'}.
}
\description{
Multidimensional vulnerability index, using truncated or extended
    niches.
}
\examples{
## Load data
#     require(vuln)
#     require(ecole)
data(braun, package='ecole')
spe <- braun$spe
env <- braun$env
## Vulnerability estimates
v <- vuln(spe, y=env$bio1) # w/o bingrid, may be 'truncated' niche
## Mapping
d <- data.frame(env, t1 = gv(v,1), t2 = gv(v,2), t3 = gv(v,3))
ecole::set_par(4)
plot(d$lat, d$lon, col=ecole::colvec(d$bio1), pch=16, main='bio1')
plot(d$lat, d$lon, col=ecole::colvec(d$t1),   pch=16, main='t1')
plot(d$lat, d$lon, col=ecole::colvec(d$t2),   pch=16, main='t2')
plot(d$lat, d$lon, col=ecole::colvec(d$t3),   pch=16, main='t3')

}
