% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuln.R
\name{vuln}
\alias{vuln}
\alias{gv}
\title{Multidimensional vulnerability index}
\usage{
vuln(spe, y, ybin, yinv, na.rm = TRUE, fullinfo = TRUE, ...)

gv(obj, pick = 1, ...)
}
\arguments{
\item{spe}{species abundance or occurrence matrix (response
variables).}

\item{y}{vector, matrix, or list of explanatory variables (climate,
fire, etc), where rows correspond to rows in \code{spe}.}

\item{ybin}{optional, data.frame or list of data.frames from
\code{bingrid}, with binned explanatory variables from
rasterizing.  Function gives extended niches if supplied,
possibly truncated niches if not supplied.}

\item{yinv}{vector of logicals, in order of \code{y}, defining
which explanatory variables y should be `inverted` (e.g., low
moisture = high vulnerability). Defaults to
\code{rep(FALSE, NCOL(y))}.}

\item{na.rm}{logical, indicating whether to strip NA values before
computation. Default is \code{na.rm = TRUE}}

\item{fullinfo}{logical, default \code{fullinfo=TRUE} propagates
any \code{NA} cases that occur in any \code{y} variable;
when \code{fullinfo=TRUE}, percentiles will be based only on
non-NA \code{y} cases.}

\item{...}{further arguments passed to other functions.}

\item{obj}{object of class \code{'vuln'}.}

\item{pick}{numeric, indicating which vulnerability metric (1,2,3).}
}
\value{
List object of class \code{'vuln'}, containing:
\describe{
  \item{\code{v}}{\code{y} values at all sites in \code{spe}.}
  \item{\code{ve}}{\code{y} values at all sites in \code{ybin}.}
  \item{\code{vo}}{\strong{vulnerable} occurrences matrix.}
  \item{\code{va}}{\strong{vulnerable} abundances matrix.}
  \item{\code{t1}}{values for 1st vulnerability index, describing
      the proportion of observed species which are vulnerable.}
  \item{\code{t2}}{values for 2nd vulnerability index, describing
      the community-mean percentile at the observed site.}
  \item{\code{t3}}{values for 3rd vulnerability index, describing
      the safety margin as the deviation of community-mean upper
      \code{y} limits from the observed \code{y} value.}
}
}
\description{
Multidimensional vulnerability index, using truncated or extended
    niches.
}
\examples{
## Load data
#     require(ecole)
data(braun, package='ecole')
spe <- braun$spe
env <- braun$env
## Vulnerability estimates
v <- vuln(spe, y=env$bio1) # w/o bingrid, may be 'truncated' niche
## Mapping
d <- data.frame(env, t1 = gv(v,1), t2 = gv(v,2), t3 = gv(v,3))
f <- function(d, j='bio1', ...){
     cc <- c("#420A6899","#65156E99","#89226A99",
             "#AD305D99","#CC424899","#E75D2E99",
             "#F7821299","#FCAA0F99","#F6D64599")
     plot(d$lat, d$lon, col=cc[cut(d[,j], breaks=9)], pch=16, main=j)
}
par(mfrow=c(2,2), bty='L', las=1)
f(d, 'bio1')
f(d, 't1')
f(d, 't2')
f(d, 't3')

}
